+++ modbus/ascii/mbascii.h
@@ -39,7 +39,7 @@ extern "C" {
 
 #if MB_ASCII_ENABLED > 0
 eMBErrorCode    eMBASCIIInit( UCHAR slaveAddress, UCHAR ucPort,
-                              ULONG ulBaudRate, eMBParity eParity );
+                              uint32_t ulBaudRate, eMBParity eParity );
 void            eMBASCIIStart( void );
 void            eMBASCIIStop( void );
